Added recipe book and code to test recipes
Recipe book is a scriptable ubject that can store recipe objects. a recipe is just an array of ingredients. the cooking script contains a couple of methods that manage a list of ingredients in "the pan". you can set the recips, add ingredients to the pan, and finish the dish. when you finish the dish the statistics of the quality of the dish are also calculated and can be used in the complete dish method in cooking.