Made switching happen.
It's very bugged at the moment. sometimes you can switch when you shouldn't be able to and sometimes you can't switch when you should be able to. aither way it is switchin, so i am pushing now, so we can combine it with stuff for tomorow, and i will fix the bug later.