It Works!
Blocks matching is detected properly. when they are detected a bunch of data is collected. this data is used to regenerate the blocks and move them up. once this is done the spaces that have moved are checked. if a new match is found  we start all over. this repeats untill no matches have been found.